requirements.txt(backend).txt
fastapi
uvicorn
torch
transformers
sentence-transformers
textblob
faiss-cpu
scikit-learn
numpy
tqdm
pydantic
python-dotenv
CORSMiddleware 
typing
json


